+++ simple-jwt-login/3rd-party/force_login.php
@@ -1,5 +1,10 @@
 <?php
 
+if (! defined('ABSPATH')) {
+    /** @phpstan-ignore-next-line  */
+    exit;
+} // Exit if accessed directly
+
 use SimpleJWTLogin\Modules\SimpleJWTLoginSettings;
 use SimpleJWTLogin\Repositories\Wordpress\WordPressRepository;
 
@@ -7,19 +12,32 @@ use SimpleJWTLogin\Repositories\Wordpress\WordPressRepository;
 add_filter(
     'rest_authentication_errors',
     function ($bypass) {
-        $currentURL  =
-            "http"
-            . (isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? "s" : "")
-            . "://" . sanitize_text_field(wp_unslash($_SERVER['HTTP_HOST']))
-            . sanitize_text_field(wp_unslash($_SERVER['REQUEST_URI']));
+        $host       = isset($_SERVER['HTTP_HOST']) ? sanitize_text_field(wp_unslash($_SERVER['HTTP_HOST'])) : '';
+        $requestUri = isset($_SERVER['REQUEST_URI']) ? sanitize_text_field(wp_unslash($_SERVER['REQUEST_URI'])) : '';
+        if (empty($host) || empty($requestUri)) {
+            return $bypass;
+        }
 
         $jwtSettings = new SimpleJWTLoginSettings(new WordPressRepository());
+        $currentURL =
+            "http"
+            . (isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? "s" : "")
+            . "://" . $host
+            . $requestUri;
 
-        $simpleJwtLoginUrl = $jwtSettings->getWordPressData()->getSiteUrl()
-               . '/?rest_route=/'
-               . $jwtSettings->getGeneralSettings()->getRouteNamespace();
-        if (strpos($currentURL, $simpleJwtLoginUrl) === 0) {
-            return true;
+        $simpleJWTLoginURLS = array(
+            "%s/?rest_route=/%s",
+            "%s/wp-json/%s",
+        );
+        foreach ($simpleJWTLoginURLS as $url) {
+            $check = sprintf(
+                $url,
+                $jwtSettings->getWordPressData()->getSiteUrl(),
+                $jwtSettings->getGeneralSettings()->getRouteNamespace(),
+            );
+            if (strpos($currentURL, $check) === 0) {
+                return true;
+            }
         }
 
         return $bypass;